Fix for LongTimeEvents
- Until now, the events had to be scheduled on the date or time that
they wanted to be active, that included the year, well, now the year is
no longer necessary to indicate it, since it will take the current year,
and so it will be repeating year after a year (we need at least to
restart the emulator once a year) and so it is not necessary to
intervene in the configuration files of each of the events.

- Two formats are admitted: dd MM yyyy-dd MM yyyy (this was the current
one) and dd MM-dd MM (this is the addition and allows the annual
repeats).