Fix for Character Raid Points
- It seems that for some strange reason the points that the players
obtain by killing raids were not added to the database or for some
reason it seems that they were in the inserted database, but after
restarting the server or by a sudden shutdown the data in the end they
were not really inserted. It has been used to rework the class to adapt
it to the future.